added profile obj changed bob score to long

+++ core/src/main/groovy/com/gamejam/model/Bob.groovy
@@ -12,7 +12,7 @@ class Bob {
     Vector2 position
     int currentLineNumber
     ArrayList<Integer> combo = new ArrayList<>()
-    int score
+    long score
     int ticketsPunched
 
     public Bob(int currentLineNumber) {
